conf/layer.conf: add possibility to add machine to BBFILES
Issue: LIN1018-5654

With commit 6ff2f4e66264de5d37fc4ee46870f6a6174029b3 a filter for
.bb and .bbappend files based on MACHINE was introduced in this layer.
This will cause problem when you create a new machine and will derive
that one on a already present machine in this layer. To solve this
we add a variable that can be set from another layer to append to
the filter function.
So that the customer can create a new additional layer with a their
new machine configuration.

Signed-off-by: Peter Bergin <peter.bergin@windriver.com>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>