nxp-ls1043: conf/layer.conf: add possibility to add machine to BBFILES
Issue: LIN1018-5654 LIN1019-4959

With the meta-freescale layer, a filter "BSP_NXP_DERIVED" for
.bb and .bbappend files based on MACHINE was introduced.
This will cause problem when you create a new machine and will derive
that one on a already present machine in this layer. To solve this
we add a variable that can be set from another layer to append to
the filter function.
So that the customer can create a new additional layer with a their
new machine configuration.

This will resolve the build warning when compiling other BSPs with all
layers download.

WARNING: No recipes available for:
	recipes-dpaa/fmlib/fmlib_git.bbappend
	recipes-extended/dpdk/dpdk_19.11-%.bbappend

Signed-off-by: Peter Bergin <peter.bergin@windriver.com>
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
[Xulin: add LIN1019-4959 defect describe.]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>