conf/layer.conf: initialize the variable BSP_NXP_LS1043_DERIVED
Issue: LIN1018-5654

with the last commit the variable is not initialized, so
it will lead to the bsp bb and bbapend layers are not
included into the compiling envronment then the compiling
will be failed. Now we set the value NULL, the customer
can override it in the local.conf.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>