nxp-ls1043: remove incomplete u-boot-qoriq configurations from this BSP
Issue: LIN1021-2964

For nxp-ls1043 BSP, because of the license issue, the whole bootloader
is not able to be built in wrlinux product directly and user need to download
the official NXP bootloader for ls1043rdba board. In this way, it doesn't
have any meaning to keep any u-boot related configurations in this BSP.
So, remove them to avoid confusing customer.

Signed-off-by: Meng Li <Meng.Li@windriver.com>